targets/function/requirements.txt
acme==2.6.0 ; python_version >= "3.7" and python_version < "4.0"
azure-common==1.1.28 ; python_version >= "3.7" and python_version < "4.0"
azure-core==1.27.1 ; python_version >= "3.7" and python_version < "4.0"
azure-identity==1.13.0 ; python_version >= "3.7" and python_version < "4.0"
azure-mgmt-core==1.4.0 ; python_version >= "3.7" and python_version < "4.0"
azure-mgmt-dns==8.1.0 ; python_version >= "3.7" and python_version < "4.0"
certbot-dns-azure==2.3.0 ; python_version >= "3.7" and python_version < "4.0"
certbot==2.6.0 ; python_version >= "3.7" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.7" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.7" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.7" and python_version < "4.0"
configargparse==1.5.3 ; python_version >= "3.7" and python_version < "4.0"
configobj==5.0.8 ; python_version >= "3.7" and python_version < "4.0"
cryptography==41.0.1 ; python_version >= "3.7" and python_version < "4.0"
distro==1.8.0 ; python_version >= "3.7" and python_version < "4.0"
idna==3.4 ; python_version >= "3.7" and python_version < "4.0"
isodate==0.6.1 ; python_version >= "3.7" and python_version < "4.0"
josepy==1.13.0 ; python_version >= "3.7" and python_version < "4.0"
msal-extensions==1.0.0 ; python_version >= "3.7" and python_version < "4.0"
msal==1.22.0 ; python_version >= "3.7" and python_version < "4.0"
parsedatetime==2.6 ; python_version >= "3.7" and python_version < "4.0"
portalocker==2.7.0 ; python_version >= "3.7" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.7" and python_version < "4.0"
pyjwt[crypto]==2.7.0 ; python_version >= "3.7" and python_version < "4.0"
pyopenssl==23.2.0 ; python_version >= "3.7" and python_version < "4.0"
pyrfc3339==1.1 ; python_version >= "3.7" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.7" and python_version < "4.0"
pywin32==306 ; python_version >= "3.7" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
pyyaml==6.0 ; python_version >= "3.7" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.7" and python_version < "4.0"
setuptools==68.0.0 ; python_version >= "3.7" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.7" and python_version < "4.0"
typing-extensions==4.6.3 ; python_version >= "3.7" and python_version < "4.0"
urllib3==2.0.3 ; python_version >= "3.7" and python_version < "4.0"
